pyproject.toml
[project]
name = "company-info-investigator"
version = "0.1.0"
description = "AI agent to investigate companies and gather information"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai-agents>=0.5.0",
    "langsmith>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
